add tests, and validation

+++ backend/src/controllers/IncidentController.js
@@ -51,11 +51,13 @@ module.exports = {
       .where('id', id)
       .select('ong_id')
       .first()
-    
-      if(incident.ong_id !== ong_id){
+    if(!incident){
+      return res.status(400).send({erorr: 'It was not possible to delete the incident'})
+    }
+    if(incident.ong_id !== ong_id){
       return res.status(401).json({error: 'Operation not permitted'})
     }
-
+  
     await connection('incidents').where('id', id).delete();
 
     return res.status(204).send();
